fix(api): update status when we update total
Ref. metriport/metriport-internal#785
